+++ kubejs/server_scripts/NTC/base/modrecipes/lazierae2/etcher.js
@@ -22,16 +22,8 @@ onEvent('recipes', event => {
           {
             time: 80,
             energy: 1500,
-            item: 'lazierae2:speculative_printed',
-            input1: 'emendatusenigmatica:silicon_gem',
-            input2: 'lazierae2:speculation_core_64',
-            input3: 'lazierae2:universal_press'
-          },
-          {
-            time: 80,
-            energy: 1500,
-            item: 'appliedenergistics2:logic_processor',
-            count: 9,
+            item: 'kubejs:logic_processor_wafer',
+            count: 1,
             input1: 'minecraft:gold_block',
             input2: 'minecraft:redstone_block',
             input3: 'kubejs:silicon_sheet'
@@ -39,8 +31,8 @@ onEvent('recipes', event => {
           {
             time: 80,
             energy: 1500,
-            item: 'appliedenergistics2:calculation_processor',
-            count: 9,
+            item: 'kubejs:calculation_processor_wafer',
+            count: 1,
             input1: 'appliedenergistics2:quartz_block',
             input2: 'minecraft:redstone_block',
             input3: 'kubejs:silicon_sheet'
@@ -48,8 +40,8 @@ onEvent('recipes', event => {
           {
             time: 80,
             energy: 1500,
-            item: 'appliedenergistics2:engineering_processor',
-            count: 9,
+            item: 'kubejs:engineering_processor_wafer',
+            count: 1,
             input1: 'minecraft:diamond_block',
             input2: 'minecraft:redstone_block',
             input3: 'kubejs:silicon_sheet'
@@ -129,6 +121,47 @@ onEvent('recipes', event => {
       ]
     };
 
+    dataItemsTwoItems = {
+      recipes: [
+          {
+            time: 80,
+            energy: 1500,
+            item: 'kubejs:1k_component_wafer',
+            count: 1,
+            input1: 'kubejs:logic_processor_wafer',
+            input2: 'kubejs:charged_redstone_sheet'
+          },
+          {
+            time: 80,
+            energy: 1500,
+            item: 'lazierae2:speculative_printed',
+            input1: 'lazierae2:universal_press',
+            input2: 'lazierae2:speculation_core_64'
+          },
+          {
+            time: 80,
+            energy: 1500,
+            item: 'kubejs:charged_redstone_sheet',
+            input1: 'minecraft:redstone_block',
+            input2: 'emendatusenigmatica:charged_certus_quartz_block'
+          },
+          {
+            time: 80,
+            energy: 1500,
+            item: 'kubejs:glassy_redstone_sheet',
+            input1: 'minecraft:redstone_block',
+            input2: 'appliedenergistics2:quartz_glass'
+          },
+          {
+            time: 80,
+            energy: 1500,
+            item: 'kubejs:glassy_glowstone_sheet',
+            input1: 'minecraft:glowstone',
+            input2: 'appliedenergistics2:quartz_glass'
+          }
+      ]
+    };
+
     dataTags.recipes.forEach((recipe) => {
 
         event.custom({
@@ -187,5 +220,30 @@ onEvent('recipes', event => {
           ]
       });
   });
+
+  dataItemsTwoItems.recipes.forEach((recipe) => {
+
+    event.custom({
+        type: 'lazierae2:etcher',
+        process_time: recipe.time,
+        energy_cost: recipe.energy,
+        output: {
+            item: recipe.item,
+            count: recipe.count
+        },
+        input: [
+          {
+            input: {
+              item: recipe.input1
+            }
+          },
+          {
+            input: {
+              item: recipe.input2
+            }
+          }
+        ]
+    });
+});
 
 })